refactor: authorative count storage resource
