+++ verified_libs/vstd_extra/src/resource/ghost_resource/count_ghost.rs
@@ -1,0 +1,443 @@
+//! Integer-based counting ghost resources.
+use crate::sum::*;
+use vstd::map::*;
+use vstd::modes::tracked_swap;
+use vstd::prelude::*;
+use vstd::resource::Loc;
+use vstd::resource::algebra::ResourceAlgebra;
+use vstd::resource::pcm::{PCM, Resource};
+
+verus! {
+
+// Integer-based counting ghost tokens which duplicate the retired int-based fractional resources.
+ghost enum FractionalCarrier<T, const TOTAL: u64> {
+    Value { v: T, n: int },
+    Empty,
+    Invalid,
+}
+
+impl<T, const TOTAL: u64> FractionalCarrier<T, TOTAL> {
+    spec fn new(v: T) -> Self {
+        FractionalCarrier::Value { v, n: TOTAL as int }
+    }
+}
+
+impl<T, const TOTAL: u64> ResourceAlgebra for FractionalCarrier<T, TOTAL> {
+    closed spec fn valid(self) -> bool {
+        match self {
+            FractionalCarrier::Invalid => false,
+            FractionalCarrier::Empty => true,
+            FractionalCarrier::Value { v: _, n } => 0 < n <= TOTAL,
+        }
+    }
+
+    closed spec fn op(a: Self, b: Self) -> Self {
+        match a {
+            FractionalCarrier::Invalid => FractionalCarrier::Invalid,
+            FractionalCarrier::Empty => b,
+            FractionalCarrier::Value { v: sv, n: sn } => match b {
+                FractionalCarrier::Invalid => FractionalCarrier::Invalid,
+                FractionalCarrier::Empty => a,
+                FractionalCarrier::Value { v: ov, n: on } => {
+                    if sv != ov {
+                        FractionalCarrier::Invalid
+                    } else if sn <= 0 || on <= 0 {
+                        FractionalCarrier::Invalid
+                    } else {
+                        FractionalCarrier::Value { v: sv, n: sn + on }
+                    }
+                },
+            },
+        }
+    }
+
+    proof fn valid_op(a: Self, b: Self) {
+    }
+
+    proof fn commutative(a: Self, b: Self) {
+    }
+
+    proof fn associative(a: Self, b: Self, c: Self) {
+    }
+}
+
+impl<T, const TOTAL: u64> PCM for FractionalCarrier<T, TOTAL> {
+    closed spec fn unit() -> Self {
+        FractionalCarrier::Empty
+    }
+
+    proof fn op_unit(self) {
+    }
+
+    proof fn unit_valid() {
+    }
+}
+
+pub tracked struct CountGhost<T, const TOTAL: u64 = 2> {
+    r: Resource<FractionalCarrier<T, TOTAL>>,
+}
+
+impl<T, const TOTAL: u64> CountGhost<T, TOTAL> {
+    #[verifier::type_invariant]
+    spec fn inv(self) -> bool {
+        self.r.value() is Value
+    }
+
+    pub closed spec fn id(self) -> Loc {
+        self.r.loc()
+    }
+
+    pub closed spec fn view(self) -> T {
+        self.r.value()->v
+    }
+
+    pub closed spec fn frac(self) -> int {
+        self.r.value()->n
+    }
+
+    pub open spec fn valid(self, id: Loc, frac: int) -> bool {
+        &&& self.id() == id
+        &&& self.frac() == frac
+    }
+
+    pub proof fn new(v: T) -> (tracked result: Self)
+        requires
+            TOTAL > 0,
+        ensures
+            result.frac() == TOTAL,
+            result@ == v,
+    {
+        let f = FractionalCarrier::<T, TOTAL>::new(v);
+        let tracked r = Resource::alloc(f);
+        Self { r }
+    }
+
+    pub proof fn agree(tracked self: &Self, tracked other: &Self)
+        requires
+            self.id() == other.id(),
+        ensures
+            self@ == other@,
+    {
+        use_type_invariant(self);
+        use_type_invariant(other);
+        let tracked joined = self.r.join_shared(&other.r);
+        joined.validate()
+    }
+
+    pub proof fn take(tracked &mut self) -> (tracked result: Self)
+        ensures
+            result == *old(self),
+    {
+        self.bounded();
+        let tracked mut mself = Self::dummy();
+        tracked_swap(self, &mut mself);
+        mself
+    }
+
+    pub proof fn split(tracked &mut self, n: int) -> (tracked result: Self)
+        requires
+            0 < n < old(self).frac(),
+        ensures
+            result.id() == final(self).id(),
+            final(self).id() == old(self).id(),
+            final(self)@ == old(self)@,
+            result@ == old(self)@,
+            final(self).frac() + result.frac() == old(self).frac(),
+            result.frac() == n,
+    {
+        self.bounded();
+        let tracked mut mself = Self::dummy();
+        tracked_swap(self, &mut mself);
+        use_type_invariant(&mself);
+        let tracked (r1, r2) = mself.r.split(
+            FractionalCarrier::Value { v: mself.r.value()->v, n: mself.r.value()->n - n },
+            FractionalCarrier::Value { v: mself.r.value()->v, n },
+        );
+        self.r = r1;
+        Self { r: r2 }
+    }
+
+    pub proof fn combine(tracked &mut self, tracked other: Self)
+        requires
+            old(self).id() == other.id(),
+        ensures
+            final(self).id() == old(self).id(),
+            final(self)@ == old(self)@,
+            final(self)@ == other@,
+            final(self).frac() == old(self).frac() + other.frac(),
+    {
+        self.bounded();
+        let tracked mut mself = Self::dummy();
+        tracked_swap(self, &mut mself);
+        use_type_invariant(&mself);
+        use_type_invariant(&other);
+        let tracked mut r = mself.r;
+        r.validate_2(&other.r);
+        *self = Self { r: r.join(other.r) };
+    }
+
+    pub proof fn update(tracked &mut self, v: T)
+        requires
+            old(self).frac() == TOTAL,
+        ensures
+            final(self).id() == old(self).id(),
+            final(self)@ == v,
+            final(self).frac() == old(self).frac(),
+    {
+        self.bounded();
+        let tracked mut mself = Self::dummy();
+        tracked_swap(self, &mut mself);
+        use_type_invariant(&mself);
+        let tracked r = mself.r;
+        let f = FractionalCarrier::<T, TOTAL>::Value { v, n: TOTAL as int };
+        *self = Self { r: r.update(f) };
+    }
+
+    pub proof fn update_with(tracked &mut self, tracked other: &mut Self, v: T)
+        requires
+            old(self).id() == old(other).id(),
+            old(self).frac() + old(other).frac() == TOTAL,
+        ensures
+            final(self).id() == old(self).id(),
+            final(other).id() == old(other).id(),
+            final(self).frac() == old(self).frac(),
+            final(other).frac() == old(other).frac(),
+            old(self)@ == old(other)@,
+            final(self)@ == v,
+            final(other)@ == v,
+    {
+        let ghost other_frac = other.frac();
+        other.bounded();
+        let tracked mut xother = Self::dummy();
+        tracked_swap(other, &mut xother);
+        self.bounded();
+        self.combine(xother);
+        self.update(v);
+        let tracked mut xother = self.split(other_frac);
+        tracked_swap(other, &mut xother);
+    }
+
+    pub proof fn bounded(tracked &self)
+        ensures
+            0 < self.frac() <= TOTAL,
+    {
+        use_type_invariant(self);
+        self.r.validate()
+    }
+
+    pub proof fn dummy() -> (tracked result: Self)
+        requires
+            TOTAL > 0,
+    {
+        Self::new(arbitrary())
+    }
+}
+
+/// A struct that stores and dispatches `CountGhost<T>`.
+/// Unlike `CountGhost`, it provides an `empty` state.
+/// It also remembers the value even it is empty.
+pub tracked struct CountGhostResource<T, const TOTAL: u64> {
+    tracked r: Option<CountGhost<T, TOTAL>>,
+    ghost snapshot: T,
+    ghost id: Loc,
+}
+
+impl<T, const TOTAL: u64> CountGhostResource<T, TOTAL> {
+    #[verifier::type_invariant]
+    closed spec fn type_inv(self) -> bool {
+        &&& TOTAL > 0
+        &&& 0 <= self.frac() <= TOTAL
+        &&& self.r is Some ==> {
+            &&& self.id == self.r->0.id()
+            &&& self.view() == self.r->0@
+        }
+    }
+
+    /// Type invariant.
+    pub open spec fn wf(self) -> bool {
+        &&& TOTAL > 0
+        &&& 0 <= self.frac() <= TOTAL
+    }
+
+    /// Whether this `CountGhostResource` is empty, i.e., has no fraction.
+    pub open spec fn is_empty(self) -> bool {
+        self.frac() == 0
+    }
+
+    /// Whether the fraction stored in this `CountGhostResource` is less than `TOTAL`.
+    pub open spec fn not_empty(self) -> bool {
+        !self.is_empty()
+    }
+
+    /// Whether this `CountGhostResource` has the full fraction, i.e., `TOTAL`.
+    pub open spec fn is_full(self) -> bool {
+        self.frac() == TOTAL
+    }
+
+    /// Returns the `CountGhost<T,TOTAL>` stored in this `CountGhostResource`.
+    pub closed spec fn storage(self) -> CountGhost<T, TOTAL> {
+        self.r->0
+    }
+
+    /// Returns the value of type `T` stored in this `CountGhostResource`.
+    pub closed spec fn view(self) -> T {
+        self.snapshot
+    }
+
+    /// The fractions stored in this `CountGhostResource`.
+    pub closed spec fn frac(self) -> int {
+        if self.r is None {
+            0int
+        } else {
+            self.storage().frac()
+        }
+    }
+
+    /// Returns the unique identifier.
+    pub closed spec fn id(self) -> Loc {
+        self.id
+    }
+
+    /// Create an arbitrary `CountGhostResource`. Useful as a placeholder.
+    pub proof fn arbitrary() -> (tracked res: Self)
+        requires
+            TOTAL > 0,
+    {
+        Self { r: None, snapshot: arbitrary(), id: arbitrary() }
+    }
+
+    /// Allocates a new `CountGhostResource` with the full fraction and the given value.
+    pub proof fn alloc(value: T) -> (tracked res: Self)
+        requires
+            TOTAL > 0,
+        ensures
+            res.not_empty(),
+            res.is_full(),
+            res@ == value,
+            res.wf(),
+    {
+        let tracked r = CountGhost::new(value);
+        Self { r: Some(r), snapshot: value, id: r.id() }
+    }
+
+    /// Splits a `CountGhost` with fraction 1.
+    pub proof fn split_one(tracked &mut self) -> (tracked res: CountGhost<T, TOTAL>)
+        requires
+            old(self).not_empty(),
+        ensures
+            final(self).id() == old(self).id(),
+            final(self).frac() + 1 == old(self).frac(),
+            final(self)@ == old(self)@,
+            res.frac() == 1,
+            res.id() == final(self).id(),
+            res@ == final(self)@,
+            old(self).frac() == 1 ==> final(self).is_empty(),
+            final(self).wf(),
+    {
+        use_type_invariant(&*self);
+        if self.frac() == 1 {
+            self.r.tracked_take()
+        } else {
+            self.r.tracked_borrow().bounded();
+            let tracked mut r = self.r.tracked_take();
+            let tracked res = r.split(1);
+            self.r = Some(r);
+            res
+        }
+    }
+
+    /// Splits a `CountGhost` with the given fraction.
+    pub proof fn split(tracked &mut self, n: int) -> (tracked res: CountGhost<T, TOTAL>)
+        requires
+            1 <= n <= old(self).frac(),
+        ensures
+            final(self).id() == old(self).id(),
+            final(self).frac() + n == old(self).frac(),
+            final(self)@ == old(self)@,
+            res.frac() == n,
+            res.id() == final(self).id(),
+            res@ == final(self)@,
+            old(self).frac() == n ==> final(self).is_empty(),
+            final(self).wf(),
+    {
+        use_type_invariant(&*self);
+        self.r.tracked_borrow().bounded();
+        if self.frac() == n {
+            self.r.tracked_take()
+        } else {
+            let tracked mut r = self.r.tracked_take();
+            let tracked res = r.split(n);
+            self.r = Some(r);
+            res
+        }
+    }
+
+    /// Combines a `CountGhost`.
+    pub proof fn combine(tracked &mut self, tracked other: CountGhost<T, TOTAL>)
+        requires
+            old(self).id() == other.id(),
+            other@ == old(self)@,
+        ensures
+            old(self).frac() + other.frac() > TOTAL ==> false,
+            old(self).frac() + other.frac() <= TOTAL ==> {
+                &&& final(self).id() == old(self).id()
+                &&& final(self)@ == old(self)@
+                &&& final(self).frac() == old(self).frac() + other.frac()
+                &&& final(self).wf()
+            },
+    {
+        if self.is_empty() {
+            other.bounded();
+            self.r = Some(other);
+        } else {
+            use_type_invariant(&*self);
+            let tracked mut r = self.r.tracked_take();
+            r.combine(other);
+            r.bounded();
+            self.r = Some(r);
+        }
+    }
+
+    /// `CountGhostResource` satisfies the type invariant.
+    pub proof fn validate(tracked &self)
+        ensures
+            self.wf(),
+    {
+        use_type_invariant(self);
+    }
+
+    pub proof fn validate_full(tracked &self)
+        requires
+            self.is_full(),
+        ensures
+            self.not_empty(),
+            self.frac() == TOTAL,
+            self.wf(),
+    {
+        use_type_invariant(self);
+    }
+
+    /// Updates the value stored in this `CountGhostResource`.
+    /// The fraction must be full before the update.
+    pub proof fn update(tracked &mut self, value: T)
+        requires
+            old(self).is_full(),
+        ensures
+            final(self).is_full(),
+            final(self)@ == value,
+            final(self).id() == old(self).id(),
+            final(self).wf(),
+    {
+        use_type_invariant(&*self);
+        let tracked mut r = self.r.tracked_take();
+        r.update(value);
+        self.snapshot = value;
+        self.r = Some(r);
+    }
+}
+
+pub type TokenResource<const TOTAL: u64> = CountGhostResource<(), TOTAL>;
+
+pub type Token<const TOTAL: u64> = CountGhost<(), TOTAL>;
+
+} // verus!
