+++ verified_libs/vstd_extra/src/resource/ghost_resource/mod.rs
@@ -1,5 +1,5 @@
 //! Ghost resources available for use in proofs.
-pub mod count;
+pub mod count_auth;
+pub mod count_ghost;
 pub mod csum;
 pub mod excl;
-pub mod tokens;
